pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "academic-figures-mcp"
version = "0.3.0"
description = "Academic figure agent harness for multi-step planning, generation, and evaluation through MCP"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
dependencies = [
    "mcp[cli]>=1.27.0",
    "httpx>=0.27.0",
    "google-genai>=1.0.0",
    "pillow>=10.0",
    "pyyaml>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.11",
    "bandit>=1.8",
    "pre-commit>=4.0",
]

[project.scripts]
afm-server = "src.presentation.server:main"
afm-run = "src.presentation.direct_run:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.setuptools.package-data]
src = ["py.typed"]

# ── Ruff ───────────────────────────────────────────────────
[tool.ruff]
target-version = "py310"
line-length = 99

[tool.ruff.lint]
select = [
    "E", "W",      # pycodestyle
    "F",            # pyflakes
    "I",            # isort
    "N",            # pep8-naming
    "UP",           # pyupgrade
    "B",            # flake8-bugbear
    "A",            # flake8-builtins
    "SIM",          # flake8-simplify
    "TCH",          # flake8-type-checking
    "RUF",          # ruff-specific
]
ignore = [
    "A003",         # shadow builtin (id, type)
    "B008",         # function call in default arg (FastMCP patterns)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# ── mypy ───────────────────────────────────────────────────
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = []

[[tool.mypy.overrides]]
module = [
    "google.*",
    "mcp.*",
    "yaml",
]
ignore_missing_imports = true

# ── pytest ─────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: requires external services (Gemini API, PubMed)",
]

# ── bandit ─────────────────────────────────────────────────
[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]
